refactor: 삭제 히스토리 목록 생성 로직 이관
* ```DeleteHistoryMaker``` -> ```DeleteHistory```
