+++ src/main/java/qna/domain/DeleteHistory.java
@@ -16,6 +16,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 @Entity
@@ -97,6 +99,15 @@ public class DeleteHistory {
                 ", deletedById=" + deletedByUser.getUserId() +
                 ", createDate=" + createDate +
                 '}';
+    }
+
+    public static List<DeleteHistory> listFromQuestion(Question question) {
+        final List<DeleteHistory> deleteHistoryList = new ArrayList<>();
+        deleteHistoryList.add(fromQuestion(question));
+        for (Answer answer : question.getAnswers()) {
+            deleteHistoryList.add(fromAnswer(answer));
+        }
+        return deleteHistoryList;
     }
 }
